// react-frontend/src/api/models/menu-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface MenuItemDTO
 */
export interface MenuItemDTO {
    /**
     * 
     * @type {number}
     * @memberof MenuItemDTO
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuItemDTO
     */
    'price'?: number;
    /**
     * 
     * @type {number}
     * @memberof MenuItemDTO
     */
    'kcal'?: number;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'editBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'imageHero'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'imageDetail'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof MenuItemDTO
     */
    'allergies'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'releaseDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'orderBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof MenuItemDTO
     */
    'updatedAt'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof MenuItemDTO
     */
    'released'?: boolean;
}