// react-frontend/src/api/models/admin-menu-item-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AdminMenuItemCreateRequest
 */
export interface AdminMenuItemCreateRequest {
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof AdminMenuItemCreateRequest
     */
    'price': number;
    /**
     * 
     * @type {number}
     * @memberof AdminMenuItemCreateRequest
     */
    'kcal'?: number;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'editBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'imageHero'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'imageDetail'?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof AdminMenuItemCreateRequest
     */
    'allergyIds'?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'releaseDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemCreateRequest
     */
    'orderBy'?: string;
}