// react-frontend/src/api/models/admin-menu-item-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AdminMenuItemUpdateRequest
 */
export interface AdminMenuItemUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof AdminMenuItemUpdateRequest
     */
    'price'?: number;
    /**
     * 
     * @type {number}
     * @memberof AdminMenuItemUpdateRequest
     */
    'kcal'?: number;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'editBy'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'imageHero'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'imageDetail'?: string;
    /**
     * 
     * @type {Array<number>}
     * @memberof AdminMenuItemUpdateRequest
     */
    'allergyIds'?: Array<number>;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'releaseDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof AdminMenuItemUpdateRequest
     */
    'orderBy'?: string;
}